addressed two size_t <-- number cast warnings

+++ vtk_export_ho.h
@@ -210,7 +210,7 @@ void vtk_export_ho
 	MultiGrid& destGrid = *destDom->grid();
 	GlobalMultiGridRefiner refiner(destGrid);
 #endif
-	size_t numRefs = ceil(log2(order));
+	size_t numRefs = (size_t) ceil(log2(order));
 	for (size_t iref = 0; iref < numRefs; ++iref)
 	{
 		try	{refiner.refine();}
